+++ back-end/helpers/analyseProjectHelpers.js
@@ -1,0 +1,17 @@
+/**
+ * Returns whether the list contains duplicates.
+ * @param arr
+ * @returns {boolean}
+ */
+export function hasDuplicates(arr) {
+    var counts = [];
+
+    for (var i = 0; i <= arr.length; i++) {
+        if (counts[arr[i]] === undefined) {
+            counts[arr[i]] = 1;
+        } else {
+            return true;
+        }
+    }
+    return false;
+}
